+++ src/forms/requirements-form/requirements-form.controller.js
@@ -1,21 +1,18 @@
 import angular from 'angular';
 
 class RequirementsFormController {
-  constructor($scope, TwRequirementsFormService) {
+  constructor($scope, TwRequirementsFormService, RequirementsService) {
     this.RequirementsFormService = TwRequirementsFormService;
+    this.RequirementsService = RequirementsService;
 
     if (!this.model) {
       this.model = {};
     }
 
-    if (this.requirements) {
-      this.RequirementsFormService.prepRequirements(this.requirements);
-    }
-
     // TODO move watches to $onChanges
     $scope.$watch('$ctrl.requirements', (newRequirements, oldRequirements) => {
       if (!angular.equals(newRequirements, oldRequirements)) {
-        this.RequirementsFormService.prepRequirements(this.requirements);
+        this.requirements = this.RequirementsService.prepRequirements(this.requirements);
 
         const oldType = this.model.type;
         const newType =
@@ -44,27 +41,6 @@ class RequirementsFormController {
     // TODO can we add asyncvalidator here? - prob not
   }
 
-  /**
-   * Perform the refreshRequirementsOnChange check on blur
-   */
-  // eslint-disable-next-line class-methods-use-this
-  onBlur(field) {
-    if (!field.refreshRequirementsOnChange) {
-      // eslint-disable-next-line no-useless-return
-      return;
-    }
-    // TODO disabled the form while we refresh requirements?
-
-    /*
-    if (false && this.onRefreshRequirements) {
-      // Should post the current model back to the requirements end
-      // point and update the requirements.
-      // TODO Can we handle this internally?
-      this.onRefreshRequirements();
-    }
-    */
-  }
-
   switchTab(newType, oldType) {
     const oldRequirementType = this.RequirementsFormService.findRequirementByType(
       oldType,
@@ -88,11 +64,18 @@ class RequirementsFormController {
       newRequirementType
     );
   }
+
+  onFieldsetRefreshRequirements() {
+    if (this.onRefreshRequirements) {
+      this.onRefreshRequirements();
+    }
+  }
 }
 
 RequirementsFormController.$inject = [
   '$scope',
-  'TwRequirementsFormService'
+  'TwRequirementsFormService',
+  'TwRequirementsService'
 ];
 
 export default RequirementsFormController;