+++ src/services/requirements/requirements.service.js
@@ -1,10 +1,14 @@
-import angular from 'angular';
-
 function RequirementsService($http) {
   this.prepRequirements = (alternatives) => {
-    alternatives.forEach((alternative) => {
-      this.prepFields(alternative.fields);
+    const preppedAlternatives = copyOf(alternatives);
+
+    preppedAlternatives.forEach((alternative) => {
+      this.prepLegacyAlternatives(alternative);
+
+      alternative.fields = this.prepFields(alternative.fields);
     });
+
+    return preppedAlternatives;
   };
 
   this.prepFields = (fields, model, validationMessages) => {
@@ -12,19 +16,11 @@ function RequirementsService($http) {
       return {};
     }
 
-    let preparedFields;
-    if (fields instanceof Array) {
-      preparedFields = {};
-      fields.forEach((field) => {
-        // If the field still has groups, we need to flatten to get the key
-        if (field.group) {
-          flattenGroup(field);
-        }
-        preparedFields[field.key] = copyOf(field);
-      });
-    } else {
-      preparedFields = copyOf(fields);
-    }
+    let preparedFields = copyOf(fields);
+
+    preparedFields = flattenFieldsWithGroups(preparedFields);
+    preparedFields = transformFieldArrayToMap(preparedFields);
+    preparedFields = transformNestedKeysToNestedSpecs(preparedFields);
 
     Object.keys(preparedFields).forEach((key) => {
       preparedFields[key] = this.prepField(preparedFields[key], model, validationMessages);
@@ -36,19 +32,124 @@ function RequirementsService($http) {
   this.prepField = (field, model, validationMessages) => {
     const preparedField = copyOf(field);
 
-    flattenGroup(preparedField);
-
     this.prepLegacyProps(preparedField);
-
     this.prepType(preparedField);
     this.prepPattern(preparedField);
     this.prepValuesAsync(preparedField, model);
-    this.prepValues(preparedField);
     this.prepValidationMessages(preparedField, validationMessages);
 
     return preparedField;
   };
 
+  /**
+   * In an older format we had an extra fieldGroup level, here we flatten that out
+   * So the inner arrays of fields within the different field groups are flattened
+   * to a single array, which is returned.
+   */
+  function flattenFieldsWithGroups(fields) {
+    if (fields instanceof Array) {
+      let flattenedFields = [];
+      fields.forEach((field) => {
+        // If we've been given a group with nested fields, break them out.
+        if (field.fields) {
+          flattenedFields = flattenedFields.concat(flattenFieldWithGroup(field, field.fields));
+        } else if (field.group) {
+          flattenedFields = flattenedFields.concat(flattenFieldWithGroup(field, field.group));
+        } else {
+          // Otherwise it's a regular field, just add it.
+          flattenedFields.push(field);
+        }
+      });
+      return flattenedFields;
+    }
+    return fields;
+  }
+
+  function flattenFieldWithGroup(field, subFields) {
+    // If first field doesn't have a label, use the one from the group
+    if (field.name && subFields.length && !subFields[0].name) {
+      subFields[0].name = field.name;
+    }
+
+    if (field.width && subFields.length && !subFields[0].width) {
+      subFields[0].width = field.width;
+    }
+
+    // If there was a tooltip at fieldGroup level move it to first field.
+    if (field.tooltip && subFields.length && !subFields[0].helpText) {
+      subFields[0].helpText = field.tooltip;
+    }
+
+    if (field.info && subFields.length && !subFields[0].helpText) {
+      subFields[0].helpText = field.info;
+    }
+
+    // If there are two parts of this group, render them side by side
+    if (subFields.length === 2) {
+      subFields.forEach((nestedField) => {
+        nestedField.width = 'md';
+      });
+    }
+
+    // If there are three parts, render the first two side by side
+    if (subFields.length === 3) {
+      subFields[0].width = 'md';
+      subFields[1].width = 'md';
+    }
+
+    return subFields;
+  }
+
+  /*
+   * Some older requirements return an array of fields, where it should be a map
+   * from the property name to the spec.  This converts arrays to maps.
+   */
+  function transformFieldArrayToMap(fields) {
+    if (fields instanceof Array) {
+      const fieldMap = {};
+      fields.forEach((field) => {
+        const key = field.key || field.name;
+        delete field.key;
+
+        fieldMap[key] = copyOf(field);
+      });
+      return fieldMap;
+    }
+    return fields;
+  }
+
+  /*
+   * Some older format return keys like 'address.city', expecting the value of
+   * city to be nested inside an address object.  This function creates a spec of
+   * type 'object', and nests such fields inside of it.  When we render we pass
+   * this object spec to a nested fieldset.
+   */
+  function transformNestedKeysToNestedSpecs(fieldMap) {
+    if (fieldMap instanceof Array) {
+      throw new Error('Expecting a map of fields, not an array');
+    }
+
+    const nestedFields = {};
+    Object.keys(fieldMap).forEach((key) => {
+      if (key.indexOf('.') > 0) {
+        // If the key contains a period we need to nest the fields in another object
+        const pathSections = key.split('.');
+        const nestedKey = pathSections[0];
+
+        // If this sub object doesn't exist yet, create it
+        if (!nestedFields[nestedKey]) {
+          nestedFields[nestedKey] = {
+            type: 'object',
+            fields: {}
+          };
+        }
+        nestedFields[nestedKey].fields[pathSections[1]] = fieldMap[key];
+      } else {
+        nestedFields[key] = fieldMap[key];
+      }
+    });
+    return nestedFields;
+  }
 
   this.prepType = (field) => {
     const type = field.type && field.type.toLowerCase && field.type.toLowerCase();
@@ -93,15 +194,32 @@ function RequirementsService($http) {
       default:
     }
 
-    if (!field.control) {
+    if (!field.control && field.type !== 'object') {
       field.control = this.getControlType(field);
     }
   };
 
-  this.prepLegacyProps = (field) => {
-    delete field.key;
+  this.prepLegacyAlternatives = (alternative) => {
+    if (!alternative.label && alternative.title) {
+      alternative.label = alternative.title;
+    }
+    if (!alternative.label) {
+      alternative.label = getNameFromType(alternative.type);
+    }
+    if (!alternative.type && alternative.name) {
+      alternative.type = alternative.name;
+    }
+    if (!alternative.description && alternative.tooltip) {
+      alternative.description = alternative.tooltip;
+    }
+    if (alternative.fieldGroups && !alternative.fields) {
+      alternative.fields = flattenFieldsWithGroups(alternative.fieldGroups);
+      delete alternative.fieldGroups;
+    }
+  };
 
-    if (field.name) {
+  this.prepLegacyProps = (field) => {
+    if (field.name && !field.title) {
       field.title = field.name;
       delete field.name;
     }
@@ -135,7 +253,48 @@ function RequirementsService($http) {
       field.values = field.valuesAllowed;
       delete field.valuesAllowed;
     }
+
+    if (field.values && field.values.map) {
+      field.values = this.prepLegacyValues(field.values);
+    }
+
+    if (field.values) {
+      // In some legacy arrays the first value is a placeholder, extract it.
+      if (field.values &&
+        field.values.length &&
+        field.values[0] &&
+        !field.values[0].value &&
+        field.values[0].label &&
+        !field.placeholder) {
+        field.placeholder = field.values[0].label;
+        field.values = field.values.slice(1);
+      }
+    }
   };
+
+  this.prepLegacyValues = values => values.map(prepLegacyValue);
+
+  function prepLegacyValue(value) {
+    if (!value.label && value.title) {
+      value.label = value.title;
+      delete value.title;
+    }
+    if (!value.label && value.name) {
+      value.label = value.name;
+      delete value.name;
+    }
+
+    if (!value.value && value.code) {
+      value.value = value.code;
+      delete value.code;
+    }
+    if (!value.value && value.key) {
+      value.value = value.key;
+      delete value.key;
+    }
+
+    return value;
+  }
 
   this.prepPattern = (field) => {
     if (field.pattern) {
@@ -173,21 +332,8 @@ function RequirementsService($http) {
       url: field.valuesAsync.url,
       data: postData || {}
     }).then((response) => {
-      field.values = response.data;
-      this.prepValues(field);
+      field.values = this.prepLegacyValues(response.data);
     });
-
-  this.prepValues = (field) => {
-    if (!angular.isArray(field.values)) {
-      return;
-    }
-    field.values.forEach((option) => {
-      option.value = option.value || option.key;
-      option.label = option.label || option.name;
-      delete option.key;
-      delete option.name;
-    });
-  };
 
   this.getParamValuesFromModel = (model, params) => {
     const data = {};
@@ -287,16 +433,20 @@ function getSelectionType(field) {
   return 'select';
 }
 
-function flattenGroup(field) {
-  if (field.group && field.group[0]) {
-    angular.extend(field, field.group[0]);
-    delete field.group;
-  }
-}
-
 function copyOf(obj) {
   // Object.assign is nicer, but lacks ie support
   return JSON.parse(JSON.stringify(obj));
+}
+
+/**
+ * Some older requirments formats do not include a label for alternatives
+ */
+function getNameFromType(tabType) {
+  if (tabType && tabType.length > 0) {
+    const tabNameWithSpaces = tabType.toLowerCase().split('_').join(' '); // String.replace method only replaces first instance
+    return tabNameWithSpaces.charAt(0).toUpperCase() + tabNameWithSpaces.slice(1);
+  }
+  return '';
 }